requirements.txt
requests
python-dotenv
sqlalchemy
fastapi
uvicorn[standard]
pydantic
